+++ extensions/registry/registry.csproj
@@ -9,6 +9,7 @@
   </PropertyGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\netstandard\ref\netstandard.csproj" />
+    <ProjectReference Include="..\accesscontrols\accesscontrols.csproj" />
     <Compile Include="registry.cs" />
   </ItemGroup>
   <Import Project="$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildThisFileDirectory), dir.targets))\dir.targets" />
